feat: bind address error handle
