+++ src/locales/pseudo/messages.po
@@ -2071,6 +2071,9 @@ msgstr ""
 msgid "No codes found to suspend trading."
 msgstr ""
 
+msgid "Failed to bind address."
+msgstr ""
+
 msgid "Buy Tax: {0}%"
 msgstr ""
 
@@ -2459,9 +2462,6 @@ msgid "Mint function"
 msgstr ""
 
 msgid "Clear"
-msgstr ""
-
-msgid "End"
 msgstr ""
 
 msgid "<0>Sent</0> an NFT to <1>{0}</1>"
