+++ src/locales/zh-Hans/messages.po
@@ -2076,6 +2076,9 @@ msgstr "切换"
 msgid "No codes found to suspend trading."
 msgstr "没有发现可以暂停交易的代码。"
 
+msgid "Failed to bind address."
+msgstr ""
+
 msgid "Buy Tax: {0}%"
 msgstr "购买税率: {0}%"
 
@@ -2465,9 +2468,6 @@ msgstr "铸造功能"
 
 msgid "Clear"
 msgstr "清空"
-
-msgid "End"
-msgstr "结束"
 
 msgid "<0>Sent</0> an NFT to <1>{0}</1>"
 msgstr "<0>发送</0> 一个 NFT 至 <1>{0}</1>"
